refactor: students 일급 콜렉션으로 리펙토링 작업
